# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for GoScraper.",
        "title": "GoScraper API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns a welcome message",
                "tags": [
                    "Home"
                ],
                "summary": "Get the home message",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/scrape/stackoverflow": {
            "get": {
                "description": "Retrieves the latest StackOverflow questions and returns them as a JSON response",
                "tags": [
                    "Scraping"
                ],
                "summary": "Scrape StackOverflow Questions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuestionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Optional, for successful responses"
                },
                "error": {
                    "description": "Optional, for error details",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "models.Question": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
        "models.QuestionsResponse": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Question"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}